sim: cStatistic's loadFromFile() made safer
- introduce _OPP_GNU_ATTRIBUTE()
- use GNU's __attribute__(format(...)) for checking freadvarsf() args
- opp_vsscanf(): more compliant to standard sscanf()
- %g -> %lg (follow opp_vsscanf change)

+++ src/common/stringutil.cc
@@ -227,7 +227,7 @@ int opp_vsscanf(const char *s, const char *fmt, va_list va)
                 fmt += 3;
             }
             else if (fmt[1] == 'g') {
-                k += sscanf(s, "%lg%n", va_arg(va, double *), &n);
+                k += sscanf(s, "%g%n", va_arg(va, float *), &n);
                 s += n;
                 fmt += 2;
             }
